Moving vst3 effect initialization to VST3Wrapper
