+++ src/CMakeLists.txt
@@ -543,12 +543,14 @@ list( APPEND SOURCES
          effects/VST3/VST3Effect.cpp
          effects/VST3/VST3EffectsModule.h
          effects/VST3/VST3EffectsModule.cpp
-         effects/VST3/VST3Utils.h
-         effects/VST3/VST3Utils.cpp
          effects/VST3/VST3OptionsDialog.h
          effects/VST3/VST3OptionsDialog.cpp
          effects/VST3/VST3ParametersWindow.h
          effects/VST3/VST3ParametersWindow.cpp
+         effects/VST3/VST3Utils.h
+         effects/VST3/VST3Utils.cpp
+         effects/VST3/VST3Wrapper.h
+         effects/VST3/VST3Wrapper.cpp
          effects/VST3/internal/ComponentHandler.h
          effects/VST3/internal/ComponentHandler.cpp
          effects/VST3/internal/ConnectionProxy.h
