+++ src/effects/VST3/VST3Utils.cpp
@@ -15,11 +15,6 @@
 #include "Base64.h"
 
 #include <wx/string.h>
-#include <wx/sizer.h>
-
-#include <pluginterfaces/vst/ivsteditcontroller.h>
-#include <pluginterfaces/vst/ivstparameterchanges.h>
-
 
 wxString VST3Utils::MakePluginPathString(const wxString& modulePath, const std::string& effectUIDString)
 {
@@ -97,85 +92,3 @@ wxString PresetsBufferStream::toString() const
    return Base64::Encode(mBuffer, mBuffer.getSize());
 }
 
-
-
-bool VST3Wrapper::FetchSettings(VST3EffectSettings& vst3Settings) const
-{
-   using namespace Steinberg;
-
-   auto processorState = owned(safenew PresetsBufferStream);
-   if (mEffectComponent->getState(processorState) == kResultOk)
-   {
-      vst3Settings.mProcessorStateStr = processorState->toString();
-   }
-   else
-   {
-      vst3Settings.mProcessorStateStr = std::nullopt;
-   }
-
-   auto controllerState = owned(safenew PresetsBufferStream);
-   if (mEditController->getState(controllerState) == kResultOk)
-   {
-      vst3Settings.mControllerStateStr = controllerState->toString();
-   }
-   else
-   {
-      vst3Settings.mControllerStateStr = std::nullopt;
-   }
-
-   return true;   
-}
-
-
-
-bool VST3Wrapper::StoreSettings(const VST3EffectSettings& vst3settings) const
-{
-   using namespace Steinberg;
-
-   // we need at least the processor state string, otherwise we can not set the EditController
-   if (!vst3settings.mProcessorStateStr)
-      return false;
-
-   auto processorState = PresetsBufferStream::fromString(*vst3settings.mProcessorStateStr);
-   if (mEffectComponent->setState(processorState) != kResultOk)
-      return false;
-
-   if (vst3settings.mControllerStateStr)
-   {
-      auto controllerState = PresetsBufferStream::fromString(*vst3settings.mControllerStateStr);
-
-      if (mEditController->setComponentState(processorState) != kResultOk ||
-         mEditController->setState(controllerState) != kResultOk)
-         return false;
-   }
-
-   return true;
-}
-
-
-bool VST3Wrapper::LoadPreset(Steinberg::IBStream* fileStream)
-{
-   using namespace Steinberg;
-
-   return Vst::PresetFile::loadPreset
-   (
-      fileStream,
-      FUID::fromTUID(mEffectClassInfo.ID().data()),
-      mEffectComponent.get(),
-      mEditController.get()
-   );
-}
-
-
-bool VST3Wrapper::SavePreset(Steinberg::IBStream* fileStream) const
-{
-   using namespace Steinberg;
-
-   return Vst::PresetFile::savePreset
-   (
-      fileStream,
-      FUID::fromTUID(mEffectClassInfo.ID().data()),
-      mEffectComponent.get(),
-      mEditController.get()
-   );
-}
